% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregate_singlepeptides.R
\name{agregate_singlepeptides}
\alias{agregate_singlepeptides}
\title{Aggregate mean intensity values of peptides mapped to the same gene for each 
ENSEMBL gene ID}
\usage{
agregate_singlepeptides(
  Qfeature,
  SV_seq,
  ProtFeatures,
  whichorder,
  names_samples = paste(sample_names, rep(1:3, each = 3), sep = "_")
)
}
\arguments{
\item{Qfeature}{QFeatures,
Data object.}

\item{SV_seq}{AAStringSet,
DNAString object.}

\item{ProtFeatures}{list,
ProtFeatures$ProtSeq us a DNAString object.}

\item{whichorder}{numeric,
to correctly reorder labels}

\item{names_samples}{character,
to specify sample names}
}
\value{
a dataframe
}
\description{
\code{agregate_singlepeptides} calculates the mean intensity values of 
peptides mapped to the same gene  to represent the intensity of the
gene-encoded protein.
}
\examples{
if(interactive()){
aggregatedWCL<-agregate_singlepeptides(QWCLpeptidesfiltered_clean,SV_seq,
 whichorder = c(1,2,4,3,6,5,7,10,9,8),
  names_samples=   paste(sample_names,rep(1:3,each=3),sep='_'))
}
}
