% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cRIC.R
\name{calculate_cRIC}
\alias{calculate_cRIC}
\title{calculate_cRIC}
\usage{
calculate_cRIC(WCL, RIC)
}
\arguments{
\item{WCL}{Dataframe
output from \code{\link{agregagte_singlepeptides}}}

\item{RIC}{Dataframe
output from \code{\link{agregagte_singlepeptides}}}
}
\value{
A data frame with log2(RIC/WCL) values
}
\description{
calculate_cRIC
}
\examples{
if(interactive()){
cRIC <- calculate_cRIC(aggregatedWCL_batch,aggregatedRIC)
}
}
