% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enigmRBP}
\alias{enigmRBP}
\title{RBP annotation data}
\format{
An object of class \code{data.frame} with 1217 rows and 6 columns.
}
\source{
\url{http://www.hentze.embl.de/public/RBDmap/}
}
\usage{
enigmRBP
}
\description{
`enigmRBP`  include many well-studied proteins with some roles in RNA 
biology
}
\keyword{datasets}
