% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_MaxQuant.R
\name{filter_MaxQuant}
\alias{filter_MaxQuant}
\title{filter_MaxQuant}
\usage{
filter_MaxQuant(proteins, tofilter, ratios = "Ratio.H.L.normalized.")
}
\arguments{
\item{proteins}{Data.frame,
Protein table originating from MaxQuant.}

\item{tofilter}{Character,
Name of the column(s) containing features to be filtered on.}

\item{ratios}{Character,
columns from proteinGroups.txt file that contain the data of interest.}
}
\value{
dataframe
}
\description{
filter_MaxQuant
}
\examples{
if(interactive()){
proteins_filtered <- filter_MaxQuant(proteins,
tofilter= c("Reverse" , "Potential.contaminant" ,"Only.identified.by.site"),
 ratios="Ratio.H.L.normalized.")
}
}
