% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_SILAC_se_parse.R
\name{make_SILAC_se_parse}
\alias{make_SILAC_se_parse}
\title{make_SILAC_se_parse
Data.frame to SummarizedExperiment object
conversion using file information and user input}
\usage{
make_SILAC_se_parse(
  proteins_unique,
  invert,
  numerator,
  denominator,
  columns_positions,
  ratios,
  chars = 13,
  length_labeling = 7,
  sep = "_",
  conditions = 3,
  bioreplicates = 3
)
}
\arguments{
\item{proteins_unique}{Data.frame,
Protein table with unique names annotated in the 'name' column
(output from \code{\link{make_unique}()}).}

\item{invert}{character,
SILAC channel to invert}

\item{numerator}{character,
condition of interest (RBP)}

\item{denominator}{character,
condition to make relative (Organic) (control)}

\item{columns_positions, }{integer
positions in the proteinGroups thaat contains Ration H/L normalised}

\item{ratios}{Character,
columns from proteinGroups.txt file that contain the data of interest.}

\item{chars}{integer,
number of letters to remove from end of names to make se}

\item{length_labeling}{integer,
number of letters to remove from beginning of names to make se}

\item{sep}{Character(1),
The separator used to parse the column header}

\item{conditions}{integer,
number of SILAC labels in the experiment}

\item{bioreplicates}{integer,
number of biologigical replicates}
}
\value{
A SummarizedExperiment object
with log2-transformed values, normalised to control and median substracted
(by column) .
}
\description{
make_SILAC_se_parse
Data.frame to SummarizedExperiment object
conversion using file information and user input
}
\examples{
if(interactive()){
# Load example
data <- doxTreatment
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+" &
data$Reverse != "+" ,]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns_positions<-str_which(colnames(cRIC_merged),"normalized\\\\.[A|B|C|18|4|M]")
ratios <- colnames(cRIC_merged)[str_which(colnames(cRIC_merged),"normalized\\\\.[A|B|C|18|4|M]")]
se <- make_SILAC_se_parse(cRIC_merged, columns_positions,
invert= c("Ratio.M.L.normalized.18_M_4","Ratio.H.L.normalized.18_M_4"), 
numerator= c("A", "B","C"),
denominator= c("M_4_18","18_M_4","4_18_M"),
ratios=ratios,
 conditions=3,
bioreplicates=3,chars = 13,length_labeling =7)
}
}
