% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_numbers.R
\name{plot_numbers}
\alias{plot_numbers}
\title{Plot protein numbers}
\usage{
plot_numbers(se, plot = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment,
Data object for which to plot protein numbers
(output from \code{\link{make_SILAC_se_parse}()}).}

\item{plot}{Logical(1),
If \code{TRUE} (default) the barplot is produced.
Otherwise (if \code{FALSE}), the data which the
barplot is based on are returned.}
}
\value{
Barplot of the number of identified proteins per sample
(generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_numbers} generates a barplot
of the number of identified proteins per sample.
}
\examples{
if(interactive()){
# Load example
data <- doxTreatment
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+" &
data$Reverse != "+" ,]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")
# Make SummarizedExperiment
ratios<-"ratios"
columns_positions<-grep(ratios, colnames(proteins_unique))
se <- make_SILAC_se_parse(data_unique, columns_positions,
invert= "ligth", numerator= "RBP",denominator= "Organic", conditions=2,
bioreplicates=3,chars = 0)
plot_numbers(se) #uses theme_RBP
}
}
