% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\title{plot_PCA}
\usage{
plot_PCA(data_filt, x = 1, y = 2, plot = TRUE, legend = "Conditions")
}
\arguments{
\item{data_filt}{SE object}

\item{x}{numberic for PCA x axis}

\item{y}{numberic for PCA y axis}

\item{plot}{logical}

\item{legend}{character name for legend}
}
\value{
a pca plot
}
\description{
plot_PCA
}
\examples{
if(interactive()){
plot_PCA(data_filt, x = 1, y = 2, plot = TRUE, legend = "Conditions")
}
}
