% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatterreplicates.R
\name{plot_scatterreplicates}
\alias{plot_scatterreplicates}
\title{plot_scatterreplicates}
\usage{
plot_scatterreplicates(
  aggregated_peptides,
  protein_1,
  protein_2,
  xlimits,
  ylimits,
  col1 = "red",
  col2 = "blue",
  repx = "hour18_1",
  repy = "hour18_2"
)
}
\arguments{
\item{aggregated_peptides}{Dataframe
output from \code{\link{agregagte_singlepeptides}}}

\item{protein_1}{Character,
Name of selected protein (1) to highlight on the plot}

\item{protein_2}{Character,
Name of selected protein (2) to highlight on the plot}

\item{xlimits}{Integer,
Sets the x-axis limits on plot.}

\item{ylimits}{Integer,
Sets the y-axis limits on plot.}

\item{col1}{Character,
Sets the color to highlight protein_1 on the plot}

\item{col2}{Character,
Sets the color to highlight protein_2 on the plot}

\item{repx}{Character,
Sets the replicate to represent on the x-axis}

\item{repy}{Character,
Sets the replicate to represent on the y-axis}
}
\value{
A scatter plot
}
\description{
plot_scatterreplicates
}
\examples{
if(interactive()){
plot_scatterreplicates(aggregatedWCL_batch,
xlimits=c(20,34), ylimits=c(20,34),
repx = "hour18_1",repy = "hour18_2")
}
}
