% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatterRIC.R
\name{plot_scatterRIC}
\alias{plot_scatterRIC}
\title{Title}
\usage{
plot_scatterRIC(tabletoplot, intensitiestoplot)
}
\arguments{
\item{tabletoplot}{Dataframe
First element of list output from \code{\link{test_moderateRIC}}}

\item{intensitiestoplot}{Matrix
Second element of l=list output from \code{\link{test_moderateRIC}}}
}
\value{
A matrix of scatterplots
}
\description{
Title
}
\examples{
if(interactive()){
test_moderateRIC(aggregatedWCL_batch)[[1]]$diff_hour18_hour4 -> tabletoplotWCL
test_moderateRIC(aggregatedWCL_batch)[[2]]$diff_hour18_hour4 ->  intensitiestoploWCL
plot_scatterRIC(tabletoplotWCL,intensitiestoploWCL)
}
}
