% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_singlepeptides.R
\name{plot_singlepeptides}
\alias{plot_singlepeptides}
\title{plot_singlepeptides}
\usage{
plot_singlepeptides(Qfeature, SV_seq, ENSG = TRUE, plot = TRUE)
}
\arguments{
\item{Qfeature}{QFeatures,
Data object.}

\item{SV_seq}{AAStringSet,
DNAString object.}

\item{ENSG}{Logical,
Whether or not peptide-mapping proteins are mapped to genes.}

\item{plot}{Logical(1),
If \code{TRUE} (default) the stacked barplot is produced.}
}
\value{
a stacked bar chart
}
\description{
plot_singlepeptides
}
\examples{
if(interactive()){
plot_singlepeptides(QWCLpeptidesfiltered_clean,SV_seq)
}
}
