% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcanoRIC.R
\name{plot_volcanoRIC}
\alias{plot_volcanoRIC}
\title{plot_volcanoRIC}
\usage{
plot_volcanoRIC(tabletoplot)
}
\arguments{
\item{tabletoplot}{Dataframe
First element of list output from \code{\link{test_moderateRIC}}}
}
\value{
A scatter plot (generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
plot_volcanoRIC
}
\examples{
if(interactive()){
test_moderateRIC(aggregatedWCL_batch)[[1]]$diff_hour18_hour4 -> tabletoplotWCL
}
}
