% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{QWCLpeptidessmall}
\alias{QWCLpeptidessmall}
\title{Qfeature mocked data}
\format{
An object of class \code{QFeatures} of length 1.
}
\usage{
QWCLpeptidessmall
}
\description{
`QWCLpeptidessmall` is a small `QFeatures` object for testing.
 SummarizedExperiment with 5 rows and 9 columns
}
\keyword{datasets}
