% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{reg_data}
\alias{reg_data}
\title{Sample dataset}
\format{A data frame with columns:
\describe{
 \item{ln_time}{A value between 1924 and 2015.}
 \item{events}{Maple products expressed as syrup, total in thousands of gallons.}
 \item{tx}{Gross value of maple products in thousands of Canadian dollars.}
 \item{c1}{c1}
 \item{c2}{c2}
 \item{c3}{c3}
 \item{x}{x}
}}
\source{
Simulated
}
\usage{
reg_data
}
\description{
Sample dataset
}
\examples{
\dontrun{
 reg_data
}
}
\keyword{datasets}
