% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_batcheffect.R
\name{remove_batcheffect}
\alias{remove_batcheffect}
\title{Perform batch effect correction}
\usage{
remove_batcheffect(
  aggregated_peptides,
  batch2 = c("A", "A", "A", "B", "B", "B", "C", "C", "C"),
  sample_names = c("hour18", "hour4", "mock")
)
}
\arguments{
\item{aggregated_peptides}{Dataframe 
output from \code{\link{agregate_singlepeptides}}}

\item{batch2}{Character,
indicates the corrected series of batches}

\item{sample_names}{Character(1),
indicates sample names}
}
\value{
dataframe correted with \code{\link[limma]{removeBatchEffect}}
}
\description{
\code{remove_batcheffect} performs batch correction given the second serie
of batches and a vector indicating the sample names.
}
\examples{
if(interactive()){
aggregatedWCL_batch<-remove_batcheffect(aggregatedWCL,batch2)
}
}
