\name{RIC-package}
\alias{RIC-package}
\alias{RIC}
\docType{package}
\title{
  Randomized Information Coefficient
}
\description{
  partial R port of Simone Romano's implementation of the Randomized Information Coefficient in Matlab.
  see: https://github.com/ialuronico/RandomizedInformationCoefficient
}
\details{
  So far, only the pairwise 
}
\author{
Loïc Thibaut, <Loic.Thibaut@gmail.com>

Maintainer: Loïc Thibaut <Loic.Thibaut@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
