% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIC.r
\name{ric}
\alias{ric}
\title{calculate RIC metrics using different methods}
\usage{
ric(data, marker_formula = events ~ tx + c1 + c2 + c3 + offset(ln_time),
  q_formula = events ~ tx + c1 + c2 + c3 + offset(ln_time),
  sample_size = 1000)
}
\arguments{
\item{data:}{dataset}

\item{marker_formula:}{formula for the marker}

\item{q_formula:}{q_formula}

\item{sample_size:}{sample size}
}
\value{
RIC metrics
}
\description{
calculate RIC metrics using different methods
}
\examples{
ric(reg_data, marker_formula=events~tx+c1+c2+c3+offset(ln_time))
ric(reg_data, marker_formula=events~tx+c1+c3+offset(ln_time))
ric(reg_data, marker_formula=events~tx+c1+offset(ln_time))
}
