% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIC.r
\name{ric_empirical}
\alias{ric_empirical}
\title{Calculates the relevant statistics (p(x), q(x), and AUCi) given a random sample of marker and corresponding treatment benefit.}
\usage{
ric_empirical(xb_data, b_bar = NULL)
}
\arguments{
\item{xb_data}{nX2 matrix with first column being the random draws from marker values and the second being an unbiased estimate of treatment benefit at that marker value}

\item{b_bar}{expected benefit of treating all v. treating no one. Should be populated with expected benefit of treatment without testing ONLY when the outcome is a policy-relevant metric that includes consequence of testing, otherwise the sample mean of benefits will be used;}
}
\value{
p(x), q(x), and AUCi
}
\description{
Calculates the relevant statistics (p(x), q(x), and AUCi) given a random sample of marker and corresponding treatment benefit.
}
