% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RBP.R
\name{scale_RBP2}
\alias{scale_RBP2}
\title{RBP scale 2}
\usage{
scale_RBP2()
}
\value{
ggplot scale
}
\description{
\code{scale_RBP} is the ggplot scale
in \code{\link{sequence_coverage}} to change colours
}
\examples{
if(interactive()){

}
}
