% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiq_cRIC.R
\name{semiq_cRIC}
\alias{semiq_cRIC}
\title{Enables the analysis of proteins with intensity values in at least one
condition.}
\usage{
semiq_cRIC(aggregated_peptides, condition, relative, updown)
}
\arguments{
\item{aggregated_peptides}{Dataframe 
output from \code{\link{agregagte_singlepeptides}}}

\item{condition}{Character,
Condition to perform semi-quantitative analysis on}

\item{relative}{Character,
Condition to perform semi-quantitative analysis against}

\item{updown}{Integer,
Threshold for FC to keep semi-quantitative values}
}
\value{
Dataframe with semi-quantitative values
}
\description{
\code{semiq_cRIC} returns a tabular data with one additional summary column
 per sample that contains the number of detectable intensity values.
}
\examples{
 semiq_cRIC(aggregatedRIC,condition = "hour4",relative = "mock" ,updown=2 )
}
