% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SV_seq}
\alias{SV_seq}
\title{AAStringSet viral annotation data}
\format{
An object of class \code{AAStringSet} of length 9.
}
\source{
\url{https://core.ac.uk/download/pdf/211022451.pdf}
}
\usage{
SV_seq
}
\description{
`SV_seq` is a AAStringSet with information about viral proteins
}
\keyword{datasets}
