% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_moderateRIC.R
\name{test_moderateRIC}
\alias{test_moderateRIC}
\title{test_moderateRIC}
\usage{
test_moderateRIC(
  aggregated_peptides,
  sample_names = c("hour18", "hour4", "mock")
)
}
\arguments{
\item{aggregated_peptides}{Dataframe
output from \code{\link{agregagte_singlepeptides}}}
}
\value{
Invisibly return a list with diff_table and diff_intensities for each timepoint
}
\description{
test_moderateRIC
}
\examples{
if(interactive()){
test_moderateRIC(aggregatedWCL_batch)[[1]]$diff_hour18_hour4 -> tabletoplotWCL
test_moderateRIC(aggregatedWCL_batch)[[2]]$diff_hour18_hour4 ->  intensitiestoploWCL
}
}
