% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_onesample.R
\name{test_onesample}
\alias{test_onesample}
\title{test_onesample and volcano plot}
\usage{
test_onesample(se, plot = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment,
Data object for which to plot protein numbers}
}
\value{
Volcano plot of significant proteins
(generated by \code{\link[ggplot2]{ggplot}})
}
\description{
test_onesample and volcano plot
}
\examples{
if(interactive()){
test_onesample(data_filt)
}
}
