% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RBP.R
\name{theme_RBP}
\alias{theme_RBP}
\title{RIC ggplot theme 1}
\usage{
theme_RBP()
}
\value{
ggplot theme
}
\description{
\code{theme_RBP} is the default ggplot theme used for plotting.
}
\examples{
if(interactive()){
# Load example
data <- doxTreatment
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+" &
data$Reverse != "+" ,]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
ratios<-"ratios"
columns_positions<-grep(ratios, colnames(proteins_unique))
se <- make_SILAC_se_parse(data_unique, columns_positions,
invert= "ligth", numerator= "RBP",denominator= "Organic", conditions=2,
bioreplicates=3,chars = 0)
plot_numbers(se) #uses theme_RBP
}
}
