% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RBP.R
\name{theme_RBP2}
\alias{theme_RBP2}
\title{RBP ggplot theme 2}
\usage{
theme_RBP2()
}
\value{
ggplot theme
}
\description{
\code{theme_RBP2} is the ggplot theme used for plotting
in \code{\link{sequence_coverage}} change axis labels and legend
}
\examples{
if(interactive()){
# Load example
data <- doxTreatment
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+" &
data$Reverse != "+" ,]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
ratios<-"ratios"
columns_positions<-grep(ratios, colnames(proteins_unique))
se <- make_SILAC_se_parse(data_unique, columns_positions,
invert= "ligth", numerator= "RBP",denominator= "Organic", conditions=2,
bioreplicates=3,chars = 0)
sequence_coverage(se,
 number=35,label1="heavy",label2="light",
 phase1="Organic", phase2="RBP") #uses theme_RBP2
}
}
