% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIC.r
\name{auci_mfc}
\alias{auci_mfc}
\title{Calculating AUCi using the method of forced choice as explained in the text (Appendix I); loops over all possible pairs and estimates the expected benefit of treating only the one with higher biomarker value}
\usage{
auci_mfc(xb_data)
}
\value{
the expected benefit of treating with higher biomarker value
}
\description{
Calculating AUCi using the method of forced choice as explained in the text (Appendix I); loops over all possible pairs and estimates the expected benefit of treating only the one with higher biomarker value
}
