% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cRIC.R
\name{calculate_cRIC}
\alias{calculate_cRIC}
\title{Estimate RNA binding activity
 
\code{calculate_cRIC} calculates log2 (RIC/WCL) changes, which are used as a
proxy of RNA binding activity.}
\usage{
calculate_cRIC(WCL, RIC)
}
\arguments{
\item{WCL}{Dataframe
output from \code{\link{agregagte_singlepeptides}}}

\item{RIC}{Dataframe
output from \code{\link{agregagte_singlepeptides}}}
}
\value{
A data frame with log2(RIC/WCL) values
}
\description{
Estimate RNA binding activity
 
\code{calculate_cRIC} calculates log2 (RIC/WCL) changes, which are used as a
proxy of RNA binding activity.
}
\examples{
if(interactive()){
cRIC <- calculate_cRIC(aggregatedWCL_batch,aggregatedRIC)
}
}
