% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapProt2Ensg.R
\name{MapProt2Ensg}
\alias{MapProt2Ensg}
\title{MapProt2Ensg}
\usage{
MapProt2Ensg(protein_list, ProtFeatures = ProtFeatures)
}
\arguments{
\item{protein_list}{list
list of proteins}

\item{ProtFeatures}{list
list of annotations. The array component of this list (ProtFeatures$GeneName)
is used in this function.}
}
\value{
a list of mapped genes
}
\description{
MapProt2Ensg
}
\examples{
if(interactive()){
ENSGid_who = MapProt2Ensg(ProtIDs_who)
}
}
