% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_batcheffect.R
\name{plot_batcheffect}
\alias{plot_batcheffect}
\title{Multidimensional scaling across replicates
 
\code{plot_batcheffect} plots the level of similarity across replicates.}
\usage{
plot_batcheffect(
  aggregated_peptides,
  sample_names = c("hour18", "hour4", "mock")
)
}
\arguments{
\item{aggregated_peptides}{Dataframe
output from \code{\link{agregagte_singlepeptides}}}

\item{sample_names}{character
indicates sample names}
}
\value{
Multidimensional scaling plot
}
\description{
Multidimensional scaling across replicates
 
\code{plot_batcheffect} plots the level of similarity across replicates.
}
\examples{
if(interactive()){
plot_batcheffect(aggregatedWCL)
}
}
