% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_singlepeptides.R
\name{plot_singlepeptides}
\alias{plot_singlepeptides}
\title{Plot the number of protein(s) to which each peptide maps to.}
\usage{
plot_singlepeptides(Qfeature, SV_seq, ProtFeatures, ENSG = TRUE, plot = TRUE)
}
\arguments{
\item{Qfeature}{QFeatures,
Data object.}

\item{SV_seq}{AAStringSet,
DNAString object.}

\item{ProtFeatures}{list,
ProtFeatures$ProtSeq us a DNAString object.}

\item{ENSG}{Logical,
Whether or not peptide-mapping proteins are mapped to genes.}

\item{plot}{Logical(1),
If \code{TRUE} (default) the stacked barplot is produced.}
}
\value{
a stacked bar chart
}
\description{
\code{plot_singlepeptides} generates a stacked barplot with of protein(s) to
which peptides map to.
}
\examples{
if(interactive()){
plot_singlepeptides(QWCLpeptidesfiltered_clean,SV_seq,ProtFeatures)
}
}
