% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIC.r
\name{ric_empirical}
\alias{ric_empirical}
\title{Calculates the relevant RIC statistics (p(x), q(x), and AUCi) given a random sample of marker and corresponding treatment benefit.}
\usage{
ric_empirical(xb_data, b_bar = NULL)
}
\arguments{
\item{xb_data}{nx2 matrix with first column being the random draws from marker values and the second being an unbiased estimate of treatment benefit at that marker value}

\item{b_bar}{expected benefit of treating all eligible patients vs. treating no one. Should be populated with expected benefit of treatment without testing ONLY when the outcome is a policy-relevant metric that includes the consequence of testing, otherwise the sample mean of benefits will be used;}
}
\value{
p(x), q(x), and AUCi
}
\description{
In clinical trials and observational studies the net treatment benefit is rarely available for each individual, as
the same person is either a control or a treatment case, and rarely both. However, if this information is available, ric_empirical can be used to calculate relevant RIC statistics.
}
