% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIC.r
\name{ric_parametric}
\alias{ric_parametric}
\title{Returns p(x) and q(x) desired points and AUCi, and local slope when parametric distribution for the joint distribution of marker value and expected treatment benefit is assumed. Note that this function does not need any data. Equations are provided in Appendix II of the paper.}
\usage{
ric_parametric(p_x = NA, mu_x, mu_b, sd_x, sd_b, rho, type)
}
\arguments{
\item{p_x}{points on the x-axis of ric (p(x)): can be scalar or vector}

\item{mu_x}{mean of  marker value}

\item{mu_b}{mean of  expected treatment benefit}

\item{sd_x}{SD of marker value}

\item{sd_b}{SD of  expected treatment benefit}

\item{rho}{correlation coefficient between marker and benefit. Note: do not use negative value as the underlying assumption (without loss of generality) is that higher marker value is associated with higher treatment benefit;}

\item{type}{normal or lognormal}
}
\value{
p(x), q(x), and AUCi
}
\description{
Returns p(x) and q(x) desired points and AUCi, and local slope when parametric distribution for the joint distribution of marker value and expected treatment benefit is assumed. Note that this function does not need any data. Equations are provided in Appendix II of the paper.
}
