% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_moderateRIC.R
\name{test_moderateRIC}
\alias{test_moderateRIC}
\title{Differential enrichment test}
\usage{
test_moderateRIC(
  aggregated_peptides,
  sample_names = c("hour18", "hour4", "mock")
)
}
\arguments{
\item{aggregated_peptides}{Dataframe
output from \code{\link{agregate_singlepeptides}}}

\item{sample_names}{character,
character strimg with experiment sample name information}
}
\value{
Invisibly return a list with diff_table and diff_intensities for each 
timepoint
}
\description{
\code{test_moderateRIC} performs a moderated t-test enrichment of the 
 aggregated mean intensity values for each protein for the different samples
}
\examples{
if(interactive()){
test_moderateRIC(aggregatedWCL_batch)[[1]]$diff_hour18_hour4 -> tabletoplotWCL
test_moderateRIC(aggregatedWCL_batch)[[2]]$diff_hour18_hour4 ->  intensitiestoploWCL
}
}
